+++ protostuff-runtime/src/test/java/io/protostuff/runtime/AbstractRuntimeCollectionSchemaTest.java
@@ -952,7 +952,8 @@ public abstract class AbstractRuntimeCollectionSchemaTest extends AbstractTest
         catch (RuntimeException e)
         {
             // expected
-            assertTrue(e.getMessage().startsWith(DecalMsg.class + " not assignable to "));
+            assertTrue(e.getMessage().endsWith(" is not assignable to " + 
+                    DecalMsg.class.getName()));
             return;
         }
         fail();